4_research/tools/requirements.txt
requests
feedparser
beautifulsoup4
markdownify
python-frontmatter
